+++ raster/src/main/scala/geotrellis/raster/Implicits.scala
@@ -17,9 +17,9 @@
 package geotrellis.raster
 
 import geotrellis.vector.Point
-import geotrellis.macros.{ NoDataMacros, TypeConversionMacros }
+import geotrellis.macros.{NoDataMacros, TypeConversionMacros}
 import geotrellis.vector._
-import geotrellis.util.MethodExtensions
+import geotrellis.util.{MethodExtensions, np}
 
 
 object Implicits extends Implicits
@@ -97,5 +97,33 @@ trait Implicits
         throw new GeoAttrsError("Cannot combine rasters with different dimensions." +
           s"${t._1.dimensions} does not match ${t._2.dimensions}")
       }
+  }
+
+  implicit class TilePercentileExtensions(tile: Tile) {
+    /**
+      * Compute percentile at the given breaks using the same algorithm as numpy
+      *
+      * https://docs.scipy.org/doc/numpy/reference/generated/numpy.percentile.html
+      * https://en.wikipedia.org/wiki/Percentile
+      *
+      * @param pctBreaks
+      * @return
+      */
+    def percentile(pctBreaks: Array[Double]): Array[Double] = {
+      np.percentile(tile.toArrayDouble.filter(isData(_)), pctBreaks)
+    }
+
+    /**
+      * Compute percentile at the given break using the same algorithm as numpy
+      *
+      * https://docs.scipy.org/doc/numpy/reference/generated/numpy.percentile.html
+      * https://en.wikipedia.org/wiki/Percentile
+      *
+      * @param pctBreak
+      * @return
+      */
+    def percentile(pctBreak: Double): Double = {
+      np.percentile(tile.toArrayDouble.filter(isData(_)), pctBreak)
+    }
   }
 }